requirements.txt
torch
torchaudio
transformers
umap-learn
numpy
soundfile
